Struct crdts::lwwreg::LWWReg

pub struct LWWReg<V, M> {
    pub val: V,
    pub marker: M,
}

LWWReg is a simple CRDT that contains an arbitrary value along with an Ord that tracks causality. It is the responsibility of the user to guarantee that the source of the causal element is monotonic. Don’t use timestamps unless you are comfortable with divergence.

M is a marker. It must grow monotonically and must be globally unique

Fields
val: V

val is the opaque element contained within this CRDT

marker: M

marker should be a monotonic value associated with this val

Implementations

impl<V: PartialEq, M: Ord> LWWReg<V, M>

pub fn update(&mut self, val: V, marker: M)

Updates value witnessed by the given marker.

use crdts::LWWReg;
let mut reg = LWWReg { val: 1, marker: 2 };

// updating with a smaller marker is a no-op
reg.update(2, 1);
assert_eq!(reg.val, 1);

// updating with larger marker succeeds
reg.update(2, 3);
assert_eq!(reg, LWWReg { val: 2, marker: 3 });

pub fn validate_update(&self, val: &V, marker: &M) -> Result<(), Validation>

An update is invalid if the marker is exactly the same as the current marker BUT the value is different:

use crdts::{lwwreg, LWWReg};
let mut reg = LWWReg { val: 1, marker: 2 };

// updating with a smaller marker is a no-op
assert_eq!(reg.validate_update(&32, &2), Err(lwwreg::Validation::ConflictingMarker));