Struct crdts::merkle_reg::MerkleReg

pub struct MerkleReg<T> { /* private fields */ }

The MerkleReg is a Register CRDT that uses the Merkle DAG structure to track the current value(s) held by this register. The roots of the Merkle DAG are the current concurrent values.

Implementations

impl<T> MerkleReg<T>

pub fn new() -> Self

Return a new instance of the MerkleReg

pub fn read(&self) -> Content<'_, T>

Read the current values held by the register

pub fn write(&self, value: T, children: BTreeSet<Hash>) -> Node<T>

Write the given value on top of the given children.

pub fn node(&self, hash: Hash) -> Option<&Node<T>>

Retrieve a node in the Merkle DAG by it’s hash.

Traverse the history of the register by pairing this method with the children of the nodes retrieved in Content::nodes().

pub fn all_nodes(&self) -> impl Iterator<Item = &Node<T>>

Iterator over all the nodes in the Merkle DAG.

pub fn children(&self, hash: Hash) -> Content<'_, T>

Returns the children of a node

pub fn parents(&self, hash: Hash) -> Content<'_, T>

Returns the parents of a node

pub fn num_nodes(&self) -> usize

Returns the number of nodes who are visible, i.e. their children have been seen.

pub fn num_orphans(&self) -> usize

Returns the number of nodes who are not visible due to missing children.