Trait crdts::ResetRemove

pub trait ResetRemove<A: Ord> {
    fn reset_remove(&mut self, clock: &VClock<A>);
}

CRDT’s are causal if they are built on top of vector clocks.

Required Methods

fn reset_remove(&mut self, clock: &VClock<A>)

Remove data that is strictly smaller than this clock