Struct crdts::vclock::VClock

pub struct VClock<A: Ord> {
    pub dots: BTreeMap<A, u64>,
}

A VClock is a standard vector clock. It contains a set of “actors” and associated counters. When a particular actor witnesses a mutation, their associated counter in a VClock is incremented. VClock is typically used as metadata for associated application data, rather than as the container for application data. VClock just tracks causality. It can tell you if something causally descends something else, or if different replicas are “concurrent” (were mutated in isolation, and need to be resolved externally).

Fields
dots: BTreeMap<A, u64>

dots is the mapping from actors to their associated counters

Implementations

impl<A: Ord> VClock<A>

pub fn new() -> Self

Returns a new VClock instance.

pub fn clone_without(&self, base_clock: &VClock<A>) -> VClock<A>where
A: Clone,

Returns a clone of self but with information that is older than given clock is forgotten

pub fn inc(&self, actor: A) -> Dot<A>where
A: Clone,

Generate Op to increment an actor’s counter.

Examples

use crdts::{VClock, CmRDT};
let mut a = VClock::new();

// `a.inc()` does not mutate the vclock!
let op = a.inc("A");
assert_eq!(a, VClock::new());

// we must apply the op to the VClock to have
// its edit take effect.
a.apply(op.clone());
assert_eq!(a.get(&"A"), 1);

// Op's can be replicated to another node and
// applied to the local state there.
let mut other_node = VClock::new();
other_node.apply(op);
assert_eq!(other_node.get(&"A"), 1);

pub fn get(&self, actor: &A) -> u64

Return the associated counter for this actor. All actors not in the vclock have an implied count of 0

pub fn dot(&self, actor: A) -> Dot<A>

Return the Dot for a given actor

pub fn concurrent(&self, other: &VClock<A>) -> bool

True if two vector clocks have diverged.

Examples

use crdts::{VClock, CmRDT};
let (mut a, mut b) = (VClock::new(), VClock::new());
a.apply(a.inc("A"));
b.apply(b.inc("B"));
assert!(a.concurrent(&b));

pub fn is_empty(&self) -> bool

Returns true if this vector clock contains nothing.

pub fn intersection(left: &VClock<A>, right: &VClock<A>) -> VClock<A>where
A: Clone,

Returns the common elements (same actor and counter) for two VClock instances.

pub fn glb(&mut self, other: &Self)

Reduces this VClock to the greatest-lower-bound of the given VClock and itsef, as an example see the following code.

use crdts::{VClock, Dot, ResetRemove, CmRDT};
let mut c = VClock::new();
c.apply(Dot::new(23, 6));
c.apply(Dot::new(89, 14));
let c2 = c.clone();

c.glb(&c2); // this is a no-op since `glb { c, c } = c`
assert_eq!(c, c2);

c.apply(Dot::new(43, 1));
assert_eq!(c.get(&43), 1);
c.glb(&c2); // should remove the 43 => 1 entry
assert_eq!(c.get(&43), 0);

pub fn iter(&self) -> impl Iterator<Item = Dot<&A>>

Returns an iterator over the dots in this vclock