Crate crdts

crdts is a library of thoroughly-tested, serializable CRDT's ported from the riak_dt library to rust.

Reexports

pub use vclock::VClock;

pub use orswot::Orswot;

pub use lwwreg::LWWReg;

pub use gcounter::GCounter;

pub use pncounter::PNCounter;

Modules

gcounter
gcounter contains the grow-only counter

lwwreg
lwwreg contains the last-write-wins register.

orswot
orswot contains the addition-biased or-set without tombstone. The orswot crate provides an implementation of the addition-biased OR-Set without tombstones (ORSWOT). Ported directly from riak_dt.

pncounter
pncounter contains the positive-negative counter

vclock
vclock contains the vector clock. The vclock crate provides a generic vector clock implementation.

Functions

from_binary
Attempts to reconstruct a type from binary.

to_binary
Dumps this type to binary.